+++ src/main/java/model/Paciente.java
@@ -1,4 +1,4 @@
-package modelclass;
+package model;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
