+++ src/main/resources/META-INF/MANIFEST.MF
@@ -1,1 +1,1 @@
-Main-Class: modeltest.Principal
+Main-Class: Principal
