+++ src/test/java/model/ConsultaTest.java
@@ -1,4 +1,4 @@
-package modelclass;
+package model;
 import org.junit.jupiter.api.Test;
 
 class ConsultaTest {
